changed to bigInteger in models

+++ app/models/tokens.py
@@ -6,10 +6,10 @@ class Token(db.Model):
 
     id = db.Column(db.Integer, primary_key=True, nullable=False)
     name = db.Column(db.String(255), nullable=False)
-    price = db.Column(db.Integer, nullable=False)
-    dailyVolume = db.Column(db.Integer, nullable=False)
-    dailyChange = db.Column(db.Integer, nullable=False)
-    marketCap = db.Column(db.Integer, nullable=False)
+    price = db.Column(db.BigInteger, nullable=False)
+    dailyVolume = db.Column(db.BigInteger, nullable=False)
+    dailyChange = db.Column(db.BigInteger, nullable=False)
+    marketCap = db.Column(db.BigInteger, nullable=False)
 
 
     def to_dict(self):
